requirements.txt
psycopg2-binary
python-dotenv
openai
pydantic
tqdm